Further simplification of ramp and flat tileset tiles
