+++ src/map-director/slopes-group-filler/InRampPropertiesLoader.cpp
@@ -18,8 +18,7 @@
 
 *****************************************************************************/
 
-#include "InRampTilesetTile.hpp"
-#include "RampTilesetTile.hpp"
+#include "InRampPropertiesLoader.hpp"
 
 namespace {
 
@@ -54,24 +53,3 @@ Orientation InRampPropertiesLoader::orientation_for
         throw InvalidArgument{"direction bad"};
     }
 }
-
-// ----------------------------------------------------------------------------
-
-void InRampTilesetTile::load
-    (const MapTilesetTile & map_tileset_tile,
-     const TilesetTileTexture & tileset_texture,
-     PlatformAssetsStrategy & platform)
-{
-    InRampPropertiesLoader loader;
-    loader.load(map_tileset_tile);
-    m_quad_tile.set_orientation(loader.elements_orientation());
-    m_quad_tile.setup(tileset_texture, loader.corner_elevations(), platform);
-}
-
-TileCornerElevations InRampTilesetTile::corner_elevations() const
-    { return m_quad_tile.corner_elevations(); }
-
-void InRampTilesetTile::make
-    (const NeighborCornerElevations &,
-     ProducableTileCallbacks & callbacks) const
-{ m_quad_tile.make(callbacks); }
