+++ src/map-director/slopes-group-filler/RampPropertiesLoader.cpp
@@ -1,0 +1,52 @@
+/******************************************************************************
+
+    GPLv3 License
+    Copyright (c) 2024 Aria Janke
+
+    This program is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+*****************************************************************************/
+
+#include "RampPropertiesLoader.hpp"
+
+namespace {
+
+using Orientation = RampPropertiesLoader::Orientation;
+
+} // end of <anonymous> namespace
+
+/* static */ void RampPropertiesLoader::instantiate_for
+    (const WithPropertiesLoader & with_loader)
+{
+    RampPropertiesLoader loader;
+    with_loader(loader);
+}
+
+/* private */ TileCornerElevations RampPropertiesLoader::elevation_offsets_for
+    (CardinalDirection direction) const
+{
+    using Cd = CardinalDirection;
+    switch (direction) {
+    case Cd::north: return TileCornerElevations{1, 1, 0, 0};
+    case Cd::east : return TileCornerElevations{1, 0, 0, 1};
+    case Cd::south: return TileCornerElevations{0, 0, 1, 1};
+    case Cd::west : return TileCornerElevations{0, 1, 1, 0};
+    default:
+        throw InvalidArgument{"direction bad"};
+    }
+}
+
+/* private */ Orientation RampPropertiesLoader::orientation_for
+    (CardinalDirection) const
+    { return Orientation::any_elements; }
